requirements.txt
Pillow>=9.0.0
cryptography>=3.0